‘邀请人数量

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/integration/GroupIntegrationInfoVo.java
@@ -22,7 +22,7 @@ public class GroupIntegrationInfoVo {
 	private Timestamp startTime;
 	private Timestamp endTime;
 	private Integer integration;
-	private Short inviteNum;
+	private Integer inviteNum;
 	private Integer inviteUser;
 	private Byte isNew;
 	private Byte isLook;
